+++ lambda_functions/ml_service/tests/unit/test_EnrichItem.py
@@ -386,3 +386,75 @@ class TestStoreTags:
         }
         ret = GetTags.get_tags_for_item(event, context, True, session)
         assert ret['body']['Tags'] == list_tags
+
+class TestPostTags:
+    def test_post_tags_for_item_1(self, monkeypatch):
+        monkeypatch.setenv("DBNAME", "Test")
+        os.environ["STAGE"] = "dev"
+
+        session = get_db_session(True, None)
+
+        # creating items
+        item = Item()
+        item.content = "https://corona-transition.org/rki-bestatigt-covid-19-sterblichkeitsrate-von-0-01-prozent-in" \
+                       "-deutschland?fbclid=IwAR2vLIkW_3EejFaeC5_wC_410uKhN_WMpWDMAcI-dF9TTsZ43MwaHeSl4n8%22 "
+        item.language = "de"
+        item = item_handler.create_item(item, True, session)
+
+        # store a fact check
+        event = {
+            "item": {
+                "id": item.id,
+                "content": item.content,
+                "language": item.language,
+            },
+            "Tags": [
+                "RKI",
+                "Covid",
+                "Corona Transition"
+            ]
+        }
+        context = ""
+        EnrichItem.store_itemtags(event, context, True, session)
+
+        event = {
+            "pathParameters": {
+                "item_id": item.id
+            }
+        }
+        response = GetTags.get_tags_for_item(event, context, True, session)
+        body = response['body']
+        # Deserialize if body is string
+        if isinstance(body, str):
+            tags = json.loads(body)['Tags']
+        else:
+            tags = body['Tags']
+        assert tags == ['RKI', 'Covid', 'Corona Transition']
+
+        event = {
+            "pathParameters": {
+                "item_id": item.id
+            },
+            "body": {
+                "tags": ['RKI', 'Covid-19']
+            }
+        }
+        response = GetTags.post_tags_for_item(event, context, True, session)
+        body = response['body']
+        # Deserialize if body is string
+        if isinstance(body, str):
+            tags_added = json.loads(body)['added tags']
+            tags_removed = json.loads(body)['removed tags']
+        else:
+            tags_added = body['added tags']
+            tags_removed = body['removed tags']
+        assert tags_added == ['Covid-19']
+        assert tags_removed == ['Covid', 'Corona Transition']
+        response = GetTags.get_tags_for_item(event, context, True, session)
+        body = response['body']
+        # Deserialize if body is string
+        if isinstance(body, str):
+            tags = json.loads(body)['Tags']
+        else:
+            tags = body['Tags']
+        assert tags == ['RKI', 'Covid-19']
